DRAW-194 refactor: exception 관련 코드 정리
