+++ app/support/exception/src/main/kotlin/com/xorker/draw/exception/ExceptionButtonType.kt
@@ -55,12 +55,10 @@ fun XorkerException.getButtons(): List<ExceptionButtonType> {
         AlreadyJoinRoomException,
         MaxRoomException,
         UnSupportedException,
-        InvalidMafiaGamePlayingPhaseStatusException,
         NotFoundWordException,
         InvalidBroadcastException,
         is InvalidMafiaPhaseException,
         InvalidRequestOnlyMyTurnException,
-        InvalidMafiaGameVotePhaseStatusException,
         InvalidRequestOtherPlayingException,
         AlreadyPlayingRoomException,
         -> buttonOk
