+++ app/support/exception/src/main/resources/i18n/messages_ko_KR.properties
@@ -3,5 +3,4 @@ exception.button.ok='확인'
 exception.button.cancel='취소'
 exception.button.close='닫기'
 exception.auth401.description='로그인을 다시 해주세요.'
-exception.auth401.description='로그인을 다시 해주세요.'
 exception.auth403.description='로그인을 다시 해주세요.'
