+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/exception/WebSocketExceptionHandler.kt
@@ -4,7 +4,6 @@ import com.xorker.draw.exception.XorkerException
 import com.xorker.draw.exception.XorkerExceptionHandler
 import com.xorker.draw.support.logging.logger
 import com.xorker.draw.websocket.message.request.RequestAction
-import com.xorker.draw.websocket.message.request.dto.ExceptionMessage
 import com.xorker.draw.websocket.parser.WebSocketResponseParser
 import org.springframework.stereotype.Component
 import org.springframework.web.socket.TextMessage
