+++ app/websocket/src/main/kotlin/com/xorker/draw/websocket/parser/WebSocketResponseParser.kt
@@ -2,7 +2,7 @@ package com.xorker.draw.websocket.parser
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.xorker.draw.websocket.SessionMessage
-import com.xorker.draw.websocket.message.request.dto.ExceptionMessage
+import com.xorker.draw.websocket.exception.ExceptionMessage
 import org.springframework.stereotype.Component
 
 @Component
